requirements.txt
openai==1.51.2
pandas==2.2.2
Jinja2==3.1.4
PyYAML==6.0.2
markdown==3.6
typer==0.12.4
rich==13.7.1
pydantic==2.9.2
pytest==8.3.2
mypy==1.11.2
ruff==0.6.9
